refactor(frontend/home): add home module
