+++ packages/frontend/src/App.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import './styles/app.css';
-import { Home, Payment, Main, Cart, User, UserProducts, UserCart } from './components/pages/';
+import { Payment, Main, Cart, User, UserProducts, UserCart } from './components/pages/';
 import Login from './components/modules/Login/Login';
 import Register from './components/modules/Register/Register';
+import Home from './components/modules/Home/Home';
 import { CartContextProvider } from './context/cart/cart';
 import ProtectedRoute from './ProtectedRoute';
