+++ packages/frontend/src/components/pages/index.ts
@@ -1,9 +1,8 @@
 import Main from './Main';
-import Home from './Home';
 import Payment from './Payment';
 import Cart from './Cart';
 import User from './User';
 import UserProducts from './UserProducts';
 import UserCart from './UserCart';
 
-export { Main, Home, Payment, Cart, User, UserProducts, UserCart };
+export { Main, Payment, Cart, User, UserProducts, UserCart };
